+++ vm/CPU/core0/InputOutput.cpp
@@ -5,11 +5,11 @@
 #include <iostream>
 using namespace std;
 
-void InputOutput::Write(int device, int format, int command, long *pkg)
+void InputOutput::Write(int device, int format, int command, double *pkg)
 {
    if(!ignore){
        Ports ports;
-       long data[6];
+       double data[6];
        for(int i = 0; i < 3; i++)
           data[i] = pkg[i];
        data[3] = device;
